Composerized extension
