+++ contao/config/autoload.ini
@@ -1,3 +1,7 @@
+;;
+; List modules which are required to be loaded beforehand
+;;
+requires[] = "core"
 
 ;;
 ; Configure what you want the autoload creator to register
