+++ contao/languages/en/modules.php
@@ -26,8 +26,7 @@
  * @license    http://opensource.org/licenses/lgpl-3.0.html
  */
 
-
 /**
- * Palettes
+ * Extendsion folder
  */
-$GLOBALS['TL_DCA']['tl_form_field']['palettes']['countryselect'] = '{type_legend},type,name,label;{fconfig_legend},mandatory,multiple,placeholder;{expert_legend:hide},value,class,accesskey,tabindex;{submit_legend},addSubmit';
+$GLOBALS['TL_LANG']['MOD']['countryselect'] = array('Country Select-Menu');
